Handler Diogo

+++ src/aurrasd.c
@@ -6,9 +6,11 @@
 #include <unistd.h>
 #include <string.h>
 #include <signal.h>
+#include <sys/wait.h>
 #include "request.h"
 #include "filters.h"
 
+int running = 1; // Variavel de status, para ver se o programa continua a receber pedidos
 
 int readConfig(char *path) {
     int fd = open(path, O_RDONLY);
@@ -33,8 +35,8 @@ int readConfig(char *path) {
 }
 
 void ctrl_c_handeler(int signum) {
-    printf("\n[DEBUG] - End of program by ctrl-C\n");
-    exit(0);
+    printf("\n[DEBUG] - Ending the program by ctrl-C\n");
+    running = 0;
 }
 
 
@@ -73,28 +75,28 @@ void loadClient(char *buffer) {
     //printf("Valores : %ld %ld %d \n", strlen(request), strlen(status), isStatus);
     if (isStatus == 6) sendStatus(path, pidClient);
     else {
-        //if (!fork()) {
+        if (!fork()) {
             Request r = createRequest(full, pidClient, path);
-            sendMessage(path,getPid(r), "Pending...\n");
-        if ( runRequest(r,path ) == -1) {
-            char clientFifo[40];
-            sprintf(clientFifo, "tmp/%d.pipe", pidClient);
-            kill(pidClient, SIGUSR1);
-            int fd = open(clientFifo, O_WRONLY);
-            char *err = "Erro!\n";
-            write(fd, err, strlen(err));
-            close(fd);
-            kill(pidClient, SIGUSR2);
-        } else {
-            kill(pidClient, SIGUSR1);
-            int fd = open(path, O_WRONLY);
-            char *msg = "Done";
-            write(fd, msg, strlen(msg));
-            close(fd);
-            kill(pidClient, SIGUSR2);
+            sendMessage(path, getPid(r), "Pending...\n");
+            if (runRequest(r, path) == -1) {
+                char clientFifo[40];
+                sprintf(clientFifo, "tmp/%d.pipe", pidClient);
+                kill(pidClient, SIGUSR1);
+                int fd = open(clientFifo, O_WRONLY);
+                char *err = "Erro!\n";
+                write(fd, err, strlen(err));
+                close(fd);
+                kill(pidClient, SIGUSR2);
+            } else {
+                kill(pidClient, SIGUSR1);
+                int fd = open(path, O_WRONLY);
+                char *msg = "Done";
+                write(fd, msg, strlen(msg));
+                close(fd);
+                kill(pidClient, SIGUSR2);
+            }
+            _exit(0);
         }
-        //    _exit(0);
-        //}
         printf("\nPassouRun\n");
     }
 }
@@ -121,7 +123,7 @@ int main(int argc, char *argv[]) {
     int fifofd;
     int bytesRead = 0;
     char buffer[1024];
-    while (1) {
+    while (running) {
         //Lê pedidos do fico central
         if ((fifofd = open(pathCentralFIfo, O_RDONLY)) < 0) {
             perror("fifo not open\n");
@@ -139,5 +141,11 @@ int main(int argc, char *argv[]) {
         printf("[DEBUG]: End of one client\n");
         close(fifofd);
     }
+
+    pid_t pid;
+    while ((pid = wait(NULL)) > 0) printf("[DEBUG] Filho com o pid %d terminou!", pid);
+
+    unlink(pathCentralFIfo);
+
     return 0;
 }
